+++ frontend/terminal/src/App.tsx
@@ -1,15 +1,15 @@
-import React from 'react';
-import './App.css';
-import Wrapper from './components/Wrapper';
-import Header from './components/Header';
-import Home from './components/Home';
-import Footer from './components/Footer';
-import { ThemeProvider } from '@mui/system';
-import theme from './components/theme';
+import React from "react";
+import "./App.css";
+import Wrapper from "./components/Wrapper";
+import Header from "./components/Header";
+import Home from "./components/Home";
+import Footer from "./components/Footer";
+import { ThemeProvider } from "@mui/system";
+import theme from "./components/theme";
 
 function App() {
-  return (
-    <div className="App">
+	return (
+		<div className="App">
     	<ThemeProvider theme={theme}>
 				<Wrapper>
 					<Header />
@@ -17,8 +17,8 @@ function App() {
 					<Footer />
 				</Wrapper>
 			</ThemeProvider>
-    </div>
-  );
+		</div>
+	);
 }
 
 export default App;