+++ frontend/terminal/src/components/Wrapper.tsx
@@ -1,0 +1,15 @@
+import React, { ReactNode } from "react";
+
+interface MyComponentProps {
+  children: ReactNode;
+}
+
+const Wrapper: React.FC<MyComponentProps> = ({ children }) => {
+	return (
+		<div className="wrapper">
+			{children}
+		</div>
+	);
+};
+
+export default Wrapper;